Extract JSDoc/doc comment functions to dedicated module
- Created src/plugin/src/transforms/feather/doc-comment-fixes.ts (830 lines)
- Removed 1,039 lines from apply-feather-fixes.ts (16,659 → 15,620)
- Extracted functions for JSDoc type sanitization, parameter resolution, and comment metadata
- Updated imports and exports to maintain functionality
- Aligned with file's own architectural recommendations
